% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googler.R
\name{search_google}
\alias{search_google}
\alias{search_google_news}
\title{search_google}
\usage{
search_google(query, start = NULL, count = NULL, news = NULL,
  tld = NULL, lang = NULL, exact = NULL, time = NULL,
  site = NULL, unfilter = NULL)

search_google_news(query, start = NULL, count = NULL, tld = NULL,
  lang = NULL, exact = NULL, time = NULL, site = NULL,
  unfilter = NULL)
}
\arguments{
\item{start}{Start at the Nth result.}

\item{count}{Show N results–defaults to 10}

\item{news}{Show results from news section.}

\item{tld}{Country-specific search with top-level domain TLD, e.g., 'in' for India.}

\item{lang}{Search for the language, e.g., 'fi' for Finnish.}

\item{exact}{Disable automatic spelling correction. Search exact keywords.}

\item{time}{Time limit search, e.g., 'h5' (5 hrs), 'd5' (5 days), 'w5' (5 weeks), 'm5' (5 months), 'y5' (5 years)}

\item{site}{Search a site using Google.}

\item{unfilter}{Do not omit similar results.}

\item{start}{Start at the Nth result.}

\item{count}{Show N results–defaults to 10}

\item{tld}{Country-specific search with top-level domain TLD, e.g., 'in' for India.}

\item{lang}{Search for the language, e.g., 'fi' for Finnish.}

\item{exact}{Disable automatic spelling correction. Search exact keywords.}

\item{time}{Time limit search, e.g., 'h5' (5 hrs), 'd5' (5 days), 'w5' (5 weeks), 'm5' (5 months), 'y5' (5 years)}

\item{site}{Search a site using Google.}

\item{unfilter}{Do not omit similar results.}
}
\description{
search_google(): Search Google from the R console

search_google_news(): Search Google NEWS from the R console
}
\examples{
\dontrun{
## search google within past month
search_google("rstats", time = "m1")

## search google news
search_google_news("rstats")
}
}
