% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googler.R
\name{googler}
\alias{googler}
\title{googler}
\usage{
googler(query = "#rstats", start = NULL, count = NULL, news = NULL,
  tld = NULL, lang = NULL, exact = NULL, time = NULL,
  site = NULL, unfilter = NULL)
}
\arguments{
\item{start}{Start at the Nth result.}

\item{count}{Show N results–defaults to 10}

\item{news}{Show results from news section.}

\item{tld}{Country-specific search with top-level domain TLD, e.g., 'in' for India.}

\item{lang}{Search for the language, e.g., 'fi' for Finnish.}

\item{exact}{Disable automatic spelling correction. Search exact keywords.}

\item{time}{Time limit search, e.g., 'h5' (5 hrs), 'd5' (5 days), 'w5' (5 weeks), 'm5' (5 months), 'y5' (5 years)}

\item{site}{Search a site using Google.}

\item{unfilter}{Do not omit similar results.}
}
\description{
Google from the R console
}
\examples{
\dontrun{
## search google within past month
googler("rstats", time = "m1")

## search google news
googler("rstats", news = TRUE)
}
}
